Take into account review comments
